# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Fleet Fuel, Idling & Utilisation Cost — API","version":"2.0.0","description":"REST API for the Fleet Intelligence Platform. All endpoints (except /token) require JWT Bearer authentication. Rate limited at 100 req/min (20 req/min for /chat)."},"paths":{"/token":{"post":{"summary":"Login","description":"Generate a JWT access token valid for 1 hour using demo credentials.","operationId":"login_token_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"Access token","content":{"application/json":{"schema":{"properties":{"access_token":{"type":"string"},"token_type":{"type":"string","example":"bearer"}}}}}},"401":{"description":"Invalid credentials"}}}},"/vehicles":{"get":{"summary":"List Vehicles","description":"List vehicles with optional filtering by fleet_id or fuel_type. Offset-paginated.","operationId":"list_vehicles","security":[{"HTTPBearer":[]}],"parameters":[{"name":"fleet_id","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"fuel_type","in":"query","required":false,"schema":{"type":"string","enum":["petrol","diesel","hybrid","ev"]}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":50}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0}}],"responses":{"200":{"description":"Vehicle list"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"}}}},"/vehicles/{vehicle_id}/cost-summary":{"get":{"summary":"Vehicle Cost Summary","description":"Daily cost breakdown (fuel + idle cost, utilisation%) for a specific vehicle.","operationId":"vehicle_cost_summary","security":[{"HTTPBearer":[]}],"parameters":[{"name":"vehicle_id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"from_date","in":"query","required":true,"schema":{"type":"string","format":"date","example":"2026-08-01"}},{"name":"to_date","in":"query","required":true,"schema":{"type":"string","format":"date","example":"2026-08-31"}}],"responses":{"200":{"description":"Daily cost rows"},"401":{"description":"Unauthorized"},"404":{"description":"Vehicle not found or no data in range"}}}},"/vehicles/{vin}/live-status":{"get":{"summary":"Live Vehicle Status","description":"Ephemeral live vehicle status from the Redis cache. Updated in real-time by the ingestion consumer as telemetry events arrive. Returns 404 if no live events have been seen for this VIN yet.","operationId":"get_live_status","security":[{"HTTPBearer":[]}],"parameters":[{"name":"vin","in":"path","required":true,"schema":{"type":"string","minLength":17,"maxLength":17,"example":"AB1CD2EF3GH4IJ567"}}],"responses":{"200":{"description":"Live status from Redis","content":{"application/json":{"schema":{"properties":{"lat":{"type":"number"},"lon":{"type":"number"},"speed_kmh":{"type":"number"},"ignition_status":{"type":"boolean"},"last_seen":{"type":"string","format":"date-time"},"sequence":{"type":"integer"}}}}}},"401":{"description":"Unauthorized"},"404":{"description":"No live data available for this VIN"},"503":{"description":"Redis cache unavailable"}}}},"/fleet/summary":{"get":{"summary":"Fleet Summary","description":"Headline monthly KPIs (total fuel cost, total idle cost, avg utilisation) from the monthly_fleet_cost materialized view.","operationId":"fleet_summary","security":[{"HTTPBearer":[]}],"parameters":[{"name":"month","in":"query","required":true,"schema":{"type":"string","format":"date","description":"First day of the month, e.g. 2026-08-01"}},{"name":"fleet_id","in":"query","required":false,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Monthly fleet cost summary"},"401":{"description":"Unauthorized"}}}},"/fleet/offenders":{"get":{"summary":"Top Offenders","description":"Top-K worst offenders ranked by a composite weighted score: normalized idle cost + idle-as-%-of-active-time + normalized idle minutes.","operationId":"top_offenders","security":[{"HTTPBearer":[]}],"parameters":[{"name":"from_date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to_date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":10}}],"responses":{"200":{"description":"Ranked offender list with weighted scores"},"401":{"description":"Unauthorized"}}}},"/chat":{"post":{"summary":"AI Agent Chat","description":"Natural-language fleet intelligence query. The agent uses tool-calling to query real fleet data and returns a grounded recommendation. Every call is audit-logged in agent_logs.","operationId":"chat_with_agent","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Agent answer with tool call trace","content":{"application/json":{"schema":{"properties":{"query":{"type":"string"},"answer":{"type":"string"},"tool_calls":{"type":"array","items":{"type":"object"}}}}}}},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded (20/min)"},"500":{"description":"LLM API key not configured or agent error"}}}}},"components":{"schemas":{"LoginRequest":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ChatRequest":{"type":"object","required":["query"],"properties":{"query":{"type":"string","example":"Which vehicles should we investigate this week?"},"from_date":{"type":"string","format":"date","default":"2026-08-28"},"to_date":{"type":"string","format":"date","default":"2026-09-26"}}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array"}},"type":"object"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array"},"msg":{"type":"string"},"type":{"type":"string"}},"type":"object","required":["loc","msg","type"]}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}